refactor: Refactor files structure
