--- packages/client/src/components/Navbar/index.ts
@@ -1,1 +1,0 @@
-export { Navbar } from './Navbar';
