--- packages/client/src/components/Navbar/types.ts
@@ -1,3 +1,0 @@
-export type TNavbarProps = {
-  isVertical?: boolean;
-};
