+++ packages/client/src/pages/Leaderboard/index.ts
@@ -1,0 +1,1 @@
+export { Leaderboard } from './Leaderboard';
