+++ packages/client/src/pages/Profile/ChangePassword/constants.ts
@@ -1,0 +1,5 @@
+export enum ErrorMessage {
+  DEFAULT_MESSAGE = 'Что-то пошло не так. Попробуйте перезагрузить страницу',
+  OLD_PASSWORD_INCORRECT = 'Старый пароль введен неверно',
+  SERVER_ERROR = 'Старый пароль введен неверно',
+}
