+++ packages/client/src/pages/Profile/ChangePassword/types.ts
@@ -2,9 +2,3 @@ export type TFormChangeUserPassword = {
   oldPassword: string;
   newPassword: string;
 };
-
-export enum ErrorMessage {
-  DEFAULT_MESSAGE = 'Что-то пошло не так. Попробуйте перезагрузить страницу',
-  OLD_PASSWORD_INCORRECT = 'Старый пароль введен неверно',
-  SERVER_ERROR = 'Старый пароль введен неверно',
-}
