+++ packages/client/src/pages/SignIn/index.ts
@@ -1,0 +1,1 @@
+export { SignIn } from './SignIn';
