+++ packages/client/src/pages/_SignUp/index.ts
@@ -1,0 +1,1 @@
+export { SignUp } from './SignUp';
